add query meta mod
