+++ query/src/context.rs
@@ -1,21 +1,16 @@
-use std::{fmt, sync::Arc};
-
+use crate::catalog::CatalogRef;
+use crate::extension::datafusion::expr::{self, func_manager::DFSessionContextFuncAdapter};
 use datafusion::{
-    catalog::catalog::CatalogProvider,
     execution::{context::SessionState, runtime_env::RuntimeEnv},
     prelude::{SessionConfig, SessionContext},
-    scheduler::Scheduler,
 };
+use std::{fmt, sync::Arc};
 use trace::warn;
-
-use crate::extension::datafusion::expr::{self, func_manager::DFSessionContextFuncAdapter};
 
 #[derive(Clone)]
 pub struct IsiphoSessionCfg {
-    exec: Arc<Scheduler>,
     session_config: SessionConfig,
     runtime: Arc<RuntimeEnv>,
-    catalog: Option<Arc<dyn CatalogProvider>>,
 }
 
 const SIZE: usize = 1000;
@@ -25,16 +20,14 @@ pub const DEFAULT_SCHEMA: &str = "public";
 pub const DEFAULT_CATALOG: &str = "cnosdb";
 
 impl IsiphoSessionCfg {
-    pub(super) fn new(exec: Arc<Scheduler>, runtime: Arc<RuntimeEnv>) -> Self {
+    pub(super) fn new(runtime: Arc<RuntimeEnv>) -> Self {
         let session_config = SessionConfig::new()
             .with_batch_size(SIZE)
             .with_information_schema(true);
 
         Self {
-            exec,
             session_config,
             runtime,
-            catalog: None,
         }
     }
 
@@ -45,14 +38,7 @@ impl IsiphoSessionCfg {
         self
     }
 
-    pub fn with_default_catalog(self, catalog: Arc<dyn CatalogProvider>) -> Self {
-        Self {
-            catalog: Some(catalog),
-            ..self
-        }
-    }
-
-    pub fn build(self) -> IsiphoSessionCtx {
+    pub fn build(self, tenant: String, catalog: CatalogRef) -> IsiphoSessionCtx {
         let state = SessionState::with_config_rt(self.session_config, self.runtime);
 
         let mut inner = SessionContext::with_state(state);
@@ -63,22 +49,15 @@ impl IsiphoSessionCfg {
             warn!("Failed to load consdb's built-in function. err: {}", e);
         };
 
-        if let Some(default_catalog) = self.catalog {
-            inner.register_catalog(DEFAULT_CATALOG, default_catalog);
-        }
+        inner.register_catalog(tenant, catalog);
 
-        IsiphoSessionCtx {
-            inner,
-            exec: Some(self.exec),
-        }
+        IsiphoSessionCtx { inner }
     }
 }
 
 #[allow(dead_code)]
-#[derive(Default)]
 pub struct IsiphoSessionCtx {
     inner: SessionContext,
-    exec: Option<Arc<Scheduler>>,
 }
 
 impl fmt::Debug for IsiphoSessionCtx {
