+++ query/src/error.rs
@@ -1,10 +1,14 @@
 use snafu::Snafu;
 
-pub type Result<T, E = Error> = std::result::Result<T, E>;
+pub type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Debug, Snafu)]
 #[snafu(visibility(pub))]
 pub enum Error {
     #[snafu(display("stmt to datafusion plan error"))]
-    DFPlanError
+    DFPlanError,
+    #[snafu(display("cannot find table name:{}", name))]
+    TableNameError { name: String },
+    #[snafu(display("cannot find database name:{}", name))]
+    DatabaseNameError { name: String },
 }
