+++ query/src/extend_sql/ast.rs
@@ -1,5 +1,5 @@
-use std::fmt;
 use sqlparser::ast::Statement;
+use std::fmt;
 
 /// Statement representations
 #[derive(Debug, PartialEq)]
@@ -18,16 +18,14 @@ pub enum ExtStatement {
     DescribeDatabase(DescribeObject),
     ShowDatabases,
     ShowTables,
-
     //todo:  insert/update/alter
-
 }
 
 #[derive(Debug, PartialEq)]
 pub struct DropObject {
     pub object_name: String,
     pub if_exist: bool,
-    pub obj_type: ObjectType
+    pub obj_type: ObjectType,
 }
 
 #[derive(Debug, PartialEq)]
@@ -37,14 +35,13 @@ pub struct DescribeObject {
 }
 
 #[derive(Debug, PartialEq)]
-pub struct DropUser{}
+pub struct DropUser {}
 #[derive(Debug, PartialEq)]
-pub struct CreateUser{}
+pub struct CreateUser {}
 #[derive(Debug, PartialEq)]
-pub struct CreateDatabase{}
+pub struct CreateDatabase {}
 #[derive(Debug, PartialEq)]
 pub struct CreateTable {}
-
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
@@ -57,8 +54,7 @@ impl fmt::Display for ObjectType {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.write_str(match self {
             ObjectType::Table => "TABLE",
-            ObjectType::Database => "DATABASE"
+            ObjectType::Database => "DATABASE",
         })
     }
 }
-
