+++ query/src/lib.rs
@@ -1,17 +1,17 @@
 extern crate core;
 
-mod catalog;
+pub mod catalog;
 mod context;
 pub mod db;
+mod error;
 mod exec;
+mod extend_sql;
 pub mod extension;
 mod function;
 mod helper;
+mod meta;
 pub mod predicate;
 pub mod schema;
 mod stream;
 mod table;
 mod tskv_exec;
-mod extend_sql;
-mod error;
-
